requirements.txt
Flask==0.12.1
pytest==3.10.1
Werkzeug==0.11.15
Flask_Cors==3.0.10
requests==2.21.0
Pillow==8.2.0
PyJWT==2.0.1
